feat: rename InviteLinksSection, MessageAlert, ProgressBar, and ShareRecordsTable components
