+++ src/app/prompts/components/shared/progressBar.tsx
