+++ src/app/prompts/components/shared/shareRecordsTable.tsx
